Fix Jest configuration
