+++ src/display.js
@@ -1,14 +1,6 @@
 import { modifyItem, deleteItem } from './index'; // eslint-disable-line
 
-const formatDate = (input) => {
-  const options = {
-    year: 'numeric', month: 'long', day: 'numeric',
-  };
-
-  const date = new Date(input);
-  const result = date.toLocaleDateString('en-US', options);
-  return result;
-};
+import formatDate from './date';
 
 const displayProjects = (projects) => {
   const projectsMain = document.querySelector('#all-projects-content');
